QPID-6055 : [Java Broker] Ensure max frame size is always less than Integer.MAX_VALUE
git-svn-id: https://svn.apache.org/repos/asf/qpid/trunk@1621498 13f79535-47bb-0310-9956-ffa450edef68

+++ qpid/java/broker-core/src/main/java/org/apache/qpid/server/model/Broker.java
@@ -79,7 +79,7 @@ public interface Broker<X extends Broker<X>> extends ConfiguredObject<X>, EventL
 
     String BROKER_FRAME_SIZE = "qpid.broker_frame_size";
     @ManagedContextDefault(name = BROKER_FRAME_SIZE)
-    long DEFAULT_FRAME_SIZE = 65535;
+    int DEFAULT_FRAME_SIZE = 65535;
 
 
     @DerivedAttribute
