+++ qpid/java/broker-plugins/amqp-0-8-protocol/src/main/java/org/apache/qpid/server/protocol/v0_8/AMQProtocolSession.java
@@ -46,7 +46,7 @@ public interface AMQProtocolSession<T extends AMQProtocolSession<T>>
 {
     long getSessionID();
 
-    void setMaxFrameSize(long frameMax);
+    void setMaxFrameSize(int frameMax);
 
     long getMaxFrameSize();
